+++ OrchestratorV1/Tests/test.deployer.ts
@@ -4,10 +4,9 @@ import * as chai from "chai";
 import * as TypeMoq from "typemoq";
 
 import * as ri from "azure-devops-node-api/interfaces/ReleaseInterfaces";
-import * as ra from "azure-devops-node-api/ReleaseApi";
 
 import { Deployer } from "../deployer";
-import { IApproveParameters, IDeployer, IOptions, IReleaseDetails, IStageApproval } from "../interfaces";
+import { IApproveParameters, IDeployer, IHelper, IReleaseDetails, IStageApproval } from "../interfaces";
 
 describe("Deployer", () => {
 
@@ -18,37 +17,12 @@ describe("Deployer", () => {
     const requesterId = "1";
 
     const projectName = "My-Project";
-    const releaseId = 1;
-    const releaseName = "My-Release";
-
     const stageId = 1;
     const stageName = "DEV";
 
-    const options: IOptions = { retryCount: 10, retryTimeout: 5000 };
-    const releaseApiMock = TypeMoq.Mock.ofType<ra.IReleaseApi>();
+    const helperMock = TypeMoq.Mock.ofType<IHelper>();
 
     const consoleLog = console.log;
-
-    it("Should get release status", async () => {
-
-        const release = {
-
-            id: releaseId,
-            name: releaseName,
-
-        } as ri.Release;
-
-        releaseApiMock.setup((x) => x.getRelease(TypeMoq.It.isAnyString(), TypeMoq.It.isAnyNumber())).returns(() => Promise.resolve(release));
-
-        const deployer: IDeployer = new Deployer(releaseApiMock.target, options);
-
-        const result = await deployer.getReleaseStatus(projectName, releaseId);
-
-        chai.expect(result).to.not.eq(null);
-        chai.expect(result.id).eq(releaseId);
-        chai.expect(result.name).eq(releaseName);
-
-    });
 
     it("Should skip stage approval", async () => {
 
@@ -91,7 +65,7 @@ describe("Deployer", () => {
 
         } as IApproveParameters;
 
-        const deployer: IDeployer = new Deployer(releaseApiMock.target, options);
+        const deployer: IDeployer = new Deployer(helperMock.target);
 
         // Hide console output
         console.log = () => { /**/ };
@@ -148,13 +122,13 @@ describe("Deployer", () => {
 
         } as IApproveParameters;
 
-        releaseApiMock.setup((x) => x.updateReleaseApproval(TypeMoq.It.isAny(), TypeMoq.It.isAnyString(), TypeMoq.It.isAnyNumber())).returns(() => Promise.resolve({
+        helperMock.setup((x) => x.updateApproval(TypeMoq.It.isAny(), TypeMoq.It.isAnyString(), TypeMoq.It.isAnyNumber())).returns(() => Promise.resolve({
 
             status: ri.ApprovalStatus.Approved,
 
         } as ri.ReleaseApproval));
 
-        const deployer: IDeployer = new Deployer(releaseApiMock.target, options);
+        const deployer: IDeployer = new Deployer(helperMock.target);
 
         // Hide console output
         console.log = () => { /**/ };
