+++ OrchestratorV1/Tests/test.helper.ts
@@ -28,7 +28,7 @@ describe("Helper", () => {
     const definitionId = 1;
     const definitionName = "My-Definition";
 
-    const releasetId = 1;
+    const releaseId = 1;
     const releaseName = "My-Release";
     const releaseStages = [ "DEV", "TEST", "PROD" ];
     const releaseEnvironments = [
@@ -103,7 +103,7 @@ describe("Helper", () => {
 
         const releaseMock = {
 
-            id: releasetId,
+            id: releaseId,
             name: releaseName,
             environments: releaseEnvironments,
 
@@ -113,10 +113,10 @@ describe("Helper", () => {
         webApiMock.setup((x) => x.getReleaseApi()).returns(() => Promise.resolve(releaseApiMock.target));
 
         const helper: IHelper = new Helper(coreApiMock.target, releaseApiMock.target, buildApiMock.target, options);
-        const result = await helper.getRelease(projectMock, releasetId, releaseStages);
+        const result = await helper.getRelease(projectMock, releaseId, releaseStages);
 
         chai.expect(result).to.not.eq(null);
-        chai.expect(result.id).eq(releasetId);
+        chai.expect(result.id).eq(releaseId);
         chai.expect(result.name).eq(releaseName);
 
     });
@@ -138,7 +138,7 @@ describe("Helper", () => {
 
         const releaseMock = {
 
-            id: releasetId,
+            id: releaseId,
             name: releaseName,
             environments: releaseEnvironments,
 
@@ -161,7 +161,7 @@ describe("Helper", () => {
         const result = await helper.createRelease(projectMock, definitionMock, detailsMock);
 
         chai.expect(result).to.not.eq(null);
-        chai.expect(result.id).eq(releasetId);
+        chai.expect(result.id).eq(releaseId);
         chai.expect(result.name).eq(releaseName);
 
     });
@@ -176,7 +176,7 @@ describe("Helper", () => {
 
         const releaseMock = {
 
-            id: releasetId,
+            id: releaseId,
             name: releaseName,
             environments: releaseEnvironments,
 
@@ -198,7 +198,7 @@ describe("Helper", () => {
         const result = await helper.findRelease(projectMock.name!, definitionId, releaseStages, filtersMock);
 
         chai.expect(result).to.not.eq(null);
-        chai.expect(result.id).eq(releasetId);
+        chai.expect(result.id).eq(releaseId);
         chai.expect(result.name).eq(releaseName);
 
     });
@@ -215,7 +215,7 @@ describe("Helper", () => {
 
         const releaseMock = {
 
-            id: releasetId,
+            id: releaseId,
             name: releaseName,
             environments: releaseEnvironments,
             tags: [ tagFilter ],
@@ -238,7 +238,7 @@ describe("Helper", () => {
         const result = await helper.findRelease(projectMock.name!, definitionId, releaseStages, filtersMock);
 
         chai.expect(result).to.not.eq(null);
-        chai.expect(result.id).eq(releasetId);
+        chai.expect(result.id).eq(releaseId);
         chai.expect(result.name).eq(releaseName);
         chai.expect(result.tags).contains(tagFilter);
 
@@ -275,7 +275,7 @@ describe("Helper", () => {
 
         const releaseMock = {
 
-            id: releasetId,
+            id: releaseId,
             name: releaseName,
             environments: releaseEnvironments,
             artifacts: [ artifactMock ],
@@ -298,7 +298,7 @@ describe("Helper", () => {
         const result = await helper.findRelease(projectMock.name!, definitionId, releaseStages, filtersMock);
 
         chai.expect(result).to.not.eq(null);
-        chai.expect(result.id).eq(releasetId);
+        chai.expect(result.id).eq(releaseId);
         chai.expect(result.name).eq(releaseName);
         chai.expect(result.artifacts).to.not.eq(null);
         chai.expect(result.artifacts).length.gt(0);
@@ -369,7 +369,7 @@ describe("Helper", () => {
 
         const releaseMock = {
 
-            id: releasetId,
+            id: releaseId,
             name: releaseName,
             environments: releaseEnvironments,
 
@@ -392,7 +392,7 @@ describe("Helper", () => {
 
         const releaseMock = {
 
-            id: releasetId,
+            id: releaseId,
             name: releaseName,
             environments: targetEnvironments,
 
@@ -403,6 +403,27 @@ describe("Helper", () => {
 
         chai.expect(result).to.not.eq(null);
         chai.expect(result).to.eq(true);
+
+    });
+
+    it("Should get release status", async () => {
+
+        const release = {
+
+            id: releaseId,
+            name: releaseName,
+
+        } as ri.Release;
+
+        releaseApiMock.setup((x) => x.getRelease(TypeMoq.It.isAnyString(), TypeMoq.It.isAnyNumber())).returns(() => Promise.resolve(release));
+
+        const helper: IHelper = new Helper(coreApiMock.target, releaseApiMock.target, buildApiMock.target, options);
+
+        const result = await helper.getReleaseStatus(projectName, releaseId);
+
+        chai.expect(result).to.not.eq(null);
+        chai.expect(result.id).eq(releaseId);
+        chai.expect(result.name).eq(releaseName);
 
     });
